modularized ch4

+++ os4/loader.c
@@ -1,6 +1,9 @@
 #include "loader.h"
-#include "defs.h"
-#include "trap.h"
+#include "proc.h"
+#include "os4_trap.h"
+#include "../utils/defs.h"
+#include "../kernel-vm/kalloc.h"
+#include "../kernel-vm/map.h"
 
 static int app_num;
 static uint64 *app_info_ptr;
@@ -26,7 +29,7 @@ void loader_init()
 pagetable_t bin_loader(uint64 start, uint64 end, struct proc *p)
 {
 	pagetable_t pg = uvmcreate();
-	if (mappages(pg, TRAPFRAME, PGSIZE, p->trapframe, PTE_R | PTE_W) < 0) {
+	if (mappages(pg, TRAPFRAME, PGSIZE, (uint64)p->trapframe, PTE_R | PTE_W) < 0) {
         panic("map trapframe fail");
     }
 	if (!PGALIGNED(start)) {
